refactoring code

+++ app/Http/Controllers/UserController.php
@@ -18,13 +18,17 @@ class UserController extends Controller
      */
     public function index()
     {
+
+        // Check authenticated user
         if (!Auth::check()) {
             return redirect('login');
         }
 
+        // fetch users list from school
         $school_id = Auth::user()->school_id;
         $users = \App\Models\User::where('school_id', $school_id)->get();
 
+        // if user role is admin, return admin page. If not, return teachers & students list page
         if (Auth::user()->role === 'ADMIN') {
             return view('users.admin', compact('users'));
         } else {
@@ -51,9 +55,18 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
+        // check authentication
+        if (!Auth::check()) {
+            return redirect('login');
+        }
+        if (Auth::user()->role != 'ADMIN') {
+            return redirect('users')->with('error', 'You dont have permission to do this action!');
+        }
         
+        // Load user model
         $new_user = new \App\Models\User;
 
+        // defining required variables
         $email = $request->get('email');
         $full_name = $request->get('full_name');
         $role = $request->get('role');
@@ -61,11 +74,13 @@ class UserController extends Controller
         $unhashed_password = mt_rand();
         $password = Hash::make($unhashed_password);
 
+        // check if email already registered
         $check_email = DB::select("select * from users where email = '".$email."'");
         if ($check_email) {
             return redirect('users')->with('error', 'User with email '.$email.' already exists! Please try another email');
         }
 
+        // Saving new user
         $new_user->email = $email;
         $new_user->username = $email;
         $new_user->password = $password;
@@ -73,6 +88,7 @@ class UserController extends Controller
         $new_user->role = $role;
         if ($new_user->save()) {
 
+            // Sending email to new user
             $data = [
                 'email' => $email,
                 'password' => $unhashed_password,
@@ -84,22 +100,22 @@ class UserController extends Controller
                $message->to( $data['email'] , $data['full_name'])->subject('Login Information')->from('khoirkamaludin@gmail.com','Smartschool');
             });
 
+            // If the new account is a teacher, load teacher model
             if ($role == 'TEACHER') {
-                $new_teacher = new \App\Models\Teacher;
-                $new_teacher->full_name = $full_name;
-                $new_teacher->user_id = $new_user->id;
-
-                if ($new_teacher->save()) {
-                    return redirect('users')->with('success', 'Teacher with email '.$email.' invited successfully! Make sure the added teacher check the login information sent to his email!');
-                }
+                $new_person = new \App\Models\Teacher;
             }
+
+            // If new account is a student, load student model
             if ($role == 'STUDENT') {
-                $new_student = new \App\Models\Student;
-                $new_student->full_name = $full_name;
-                $new_student->user_id = $new_user->id;
-                if ($new_student->save()) {
-                    return redirect('users')->with('success', 'Student with email '.$email.' invited successfully! Make sure the added student check the login information sent to his email!');
-                }
+                $new_person = new \App\Models\Student;
+            }
+
+            // Save personal data to db
+            $new_person->full_name = $full_name;
+            $new_person->user_id = $new_user->id;
+
+            if ($new_person->save()) {
+                return redirect('users')->with('success', 'Teacher with email '.$email.' invited successfully! Make sure the added teacher check the login information sent to his email!');
             }
         }
     }
@@ -135,12 +151,32 @@ class UserController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $user = \App\Models\User::findOrFail($id);
-        if ($user->role == 'ADMIN') {
-            $school_admin = $user->school_admin;
-            $school_admin->full_name = $request->get('full_name');
-            $school_admin->save();
+
+        // Check authenticated user
+        if (!Auth::check()) {
+            return redirect('login');
         }
+        if (Auth::user()->role != 'ADMIN') {
+            return redirect('users')->with('error', 'You dont have permission to do this action!');
+        }
+
+        // Find user
+        $user = \App\Models\User::findOrFail($id);
+
+        // Edit personal data depends on user role
+        if ($user->role == 'ADMIN') {
+            $person = $user->school_admin;
+        }
+        if ($user->role == 'TEACHER') {
+            $person = $user->teacher;
+        }
+        if ($user->role == 'STUDENT') {
+            $person = $user->student;
+        }
+        $person->full_name = $request->get('full_name');
+        $person->save();
+
+        // Hash password and save
         $user->password = Hash::make($request->get('password'));
         if ($user->save()) {
             return redirect('users')->with('success', 'User '.$user->username.' updated successfully');
@@ -155,8 +191,18 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+
+        // Check authenticated user
+        if (!Auth::check()) {
+            return redirect('login');
+        }
+        if (Auth::user()->role != 'ADMIN') {
+            return redirect('users')->with('error', 'You dont have permission to do this action!');
+        }
+        // Find user
         $user = \App\Models\User::findOrFail($id);
         
+        // Delete user data and determine user role
         if ($user->delete()) {
             if ($user->role == 'ADMIN') {
                 $user_role = $user->school_admin;
@@ -168,6 +214,7 @@ class UserController extends Controller
                 $user_role = $user->student;
             }
 
+            //  Delete personal data
             if ($user_role->delete()) {
                 return redirect('users')->with('success', 'User '.$user->username.' deleted successfully');
             }
